@@ -32,26 +32,27 @@ include = [
 ]
 
 [tool.poetry.dependencies]
-python = "~2.7 || ^3.5"
-future = "*"
-jpype1 = "<=0.7.0,>=0.6.0"
-pandas = {version = ">=0.19.0", optional = true}
+# https://github.com/pandas-dev/pandas/pull/29212#issuecomment-551370118
+python = "^3.6.1"
+jpype1 = "<2.0.0,>=1.1.0"
+pandas = {version = ">=1.0.0", optional = true}
 sqlalchemy = {version = "<2.0.0,>=1.0.0", optional = true}
 
 [tool.poetry.dev-dependencies]
-colorama = "0.4.1"
 awscli = "*"
 tox = "*"
 wheel = "*"
 twine = "*"
-pandas = ">=0.19.0"
+pandas = ">=1.0.0"
 sqlalchemy = ">=1.0.0, <2.0.0"
+mypy = "*"
 pytest = ">=3.5"
 pytest-cov = "*"
 pytest-flake8 = ">=1.0.1"
 pytest-black = "*"
 pytest-isort = "*"
 pytest-xdist = "*"
+pytest-mypy = "*"
 codecov = "*"
 
 [tool.poetry.extras]
