@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, unicode_literals
-
 import logging
 import os
+from typing import Any, List, Optional
 
 import jpype
-from future.utils import iteritems
 
 from pyathenajdbc import (
     ATHENA_CONNECTION_STRING,
@@ -13,32 +11,34 @@ from pyathenajdbc import (
     ATHENA_JAR,
     LOG4J_PROPERTIES,
 )
-from pyathenajdbc.converter import JDBCTypeConverter
+from pyathenajdbc.converter import DefaultJDBCTypeConverter, JDBCTypeConverter
 from pyathenajdbc.cursor import Cursor
 from pyathenajdbc.error import NotSupportedError, ProgrammingError
-from pyathenajdbc.formatter import ParameterFormatter
+from pyathenajdbc.formatter import DefaultParameterFormatter, Formatter
 from pyathenajdbc.util import attach_thread_to_jvm, synchronized
 
-_logger = logging.getLogger(__name__)
+_logger = logging.getLogger(__name__)  # type: ignore
 
 
 class Connection(object):
 
-    _ENV_S3_STAGING_DIR = "AWS_ATHENA_S3_STAGING_DIR"
-    _ENV_S3_OUTPUT_LOCATION = "AWS_ATHENA_S3_OUTPUT_LOCATION"
-    _ENV_WORK_GROUP = "AWS_ATHENA_WORK_GROUP"
-    _BASE_PATH = os.path.dirname(os.path.abspath(__file__))
+    _ENV_S3_STAGING_DIR: str = "AWS_ATHENA_S3_STAGING_DIR"
+    _ENV_S3_OUTPUT_LOCATION: str = "AWS_ATHENA_S3_OUTPUT_LOCATION"
+    _ENV_WORK_GROUP: str = "AWS_ATHENA_WORK_GROUP"
+    _BASE_PATH: str = os.path.dirname(os.path.abspath(__file__))
+
+    _class_loader = None
 
     def __init__(
         self,
-        jvm_path=None,
-        jvm_options=None,
-        converter=None,
-        formatter=None,
-        driver_path=None,
-        log4j_conf=None,
+        jvm_path: Optional[str] = None,
+        jvm_options: Optional[List[str]] = None,
+        converter: Optional[JDBCTypeConverter] = None,
+        formatter: Optional[Formatter] = None,
+        driver_path: Optional[str] = None,
+        log4j_conf: Optional[str] = None,
         **driver_kwargs
-    ):
+    ) -> None:
         self._start_jvm(jvm_path, jvm_options, driver_path, log4j_conf)
         self._driver_kwargs = driver_kwargs
         self.region_name = self._driver_kwargs.get(
@@ -56,14 +56,20 @@ class Connection(object):
             )
         else:
             self._jdbc_conn = jpype.java.sql.DriverManager.getConnection()
-        self._converter = converter if converter else JDBCTypeConverter()
-        self._formatter = formatter if formatter else ParameterFormatter()
+        self._converter = converter if converter else DefaultJDBCTypeConverter()
+        self._formatter = formatter if formatter else DefaultParameterFormatter()
 
     @classmethod
     @synchronized
-    def _start_jvm(cls, jvm_path, jvm_options, driver_path, log4j_conf):
+    def _start_jvm(
+        cls,
+        jvm_path: Optional[str],
+        jvm_options: Optional[List[str]],
+        driver_path: Optional[str],
+        log4j_conf: Optional[str],
+    ) -> None:
         if jvm_path is None:
-            jvm_path = jpype.get_default_jvm_path()
+            jvm_path = jpype.getDefaultJVMPath()
         if driver_path is None:
             driver_path = os.path.join(cls._BASE_PATH, ATHENA_JAR)
         if log4j_conf is None:
@@ -78,28 +84,25 @@ class Connection(object):
             if jvm_options:
                 args.extend(jvm_options)
             _logger.debug("JVM args: %s", args)
-            if jpype.__version__.startswith("0.6"):
-                jpype.startJVM(jvm_path, *args)
-            else:
-                jpype.startJVM(
-                    jvm_path, *args, ignoreUnrecognized=True, convertStrings=True
-                )
-            cls.class_loader = (
+            jpype.startJVM(
+                jvm_path, *args, ignoreUnrecognized=True, convertStrings=True
+            )
+            cls._class_loader = (
                 jpype.java.lang.Thread.currentThread().getContextClassLoader()
             )
-        if not jpype.isThreadAttachedToJVM():
-            jpype.attachThreadToJVM()
-            if not cls.class_loader:
-                cls.class_loader = (
+        if not jpype.java.lang.Thread.isAttached():
+            jpype.java.lang.Thread.attach()
+            if not cls._class_loader:
+                cls._class_loader = (
                     jpype.java.lang.Thread.currentThread().getContextClassLoader()
                 )
             class_loader = jpype.java.net.URLClassLoader.newInstance(
                 [jpype.java.net.URL("jar:file:{0}!/".format(driver_path))],
-                cls.class_loader,
+                cls._class_loader,
             )
             jpype.java.lang.Thread.currentThread().setContextClassLoader(class_loader)
 
-    def _build_driver_args(self):
+    def _build_driver_args(self) -> Any:
         props = jpype.java.util.Properties()
         props.setProperty(
             "AwsCredentialsProviderClass",
@@ -117,7 +120,7 @@ class Connection(object):
             props.setProperty("Schema", self.schema_name)
         if self.work_group:
             props.setProperty("Workgroup", self.work_group)
-        for k, v in iteritems(self._driver_kwargs):
+        for k, v in self._driver_kwargs.items():
             if k and v:
                 props.setProperty(k, v)
         return props
@@ -129,28 +132,28 @@ class Connection(object):
         self.close()
 
     @attach_thread_to_jvm
-    def cursor(self):
+    def cursor(self) -> Cursor:
         if self.is_closed:
             raise ProgrammingError("Connection is closed.")
         return Cursor(self._jdbc_conn, self._converter, self._formatter)
 
     @attach_thread_to_jvm
     @synchronized
-    def close(self):
+    def close(self) -> None:
         if not self.is_closed:
             self._jdbc_conn.close()
             self._jdbc_conn = None
 
-    @property
+    @property  # type: ignore
     @attach_thread_to_jvm
-    def is_closed(self):
+    def is_closed(self) -> bool:
         return self._jdbc_conn is None or self._jdbc_conn.isClosed()
 
-    def commit(self):
+    def commit(self) -> None:
         """Athena JDBC connection is only supported for auto-commit mode."""
         pass
 
-    def rollback(self):
+    def rollback(self) -> None:
         raise NotSupportedError(
             "Athena JDBC connection is only supported for auto-commit mode."
         )
