@@ -1,45 +1,60 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, unicode_literals
-
 import logging
+from typing import Any, Dict, List, Optional, Tuple, cast
 
-from future.utils import raise_from
-from past.builtins.misc import xrange
-
+from pyathenajdbc.converter import JDBCTypeConverter
 from pyathenajdbc.error import DatabaseError, ProgrammingError
+from pyathenajdbc.formatter import Formatter
 from pyathenajdbc.util import attach_thread_to_jvm, synchronized
 
-_logger = logging.getLogger(__name__)
+_logger = logging.getLogger(__name__)  # type: ignore
 
 
 class Cursor(object):
 
-    DEFAULT_FETCH_SIZE = 1000
+    DEFAULT_FETCH_SIZE: int = 1000
 
-    def __init__(self, connection, converter, formatter):
+    def __init__(
+        self,
+        connection: Any,
+        converter: JDBCTypeConverter,
+        formatter: Formatter,
+    ):
         self._connection = connection
         self._converter = converter
         self._formatter = formatter
 
-        self._rownumber = None
-        self._arraysize = self.DEFAULT_FETCH_SIZE
+        self._rownumber: Optional[int] = None
+        self._arraysize: int = self.DEFAULT_FETCH_SIZE
 
-        self._description = None
-        self._statement = self.connection.createStatement()
-        self._result_set = None
-        self._meta_data = None
-        self._update_count = -1
+        self._description: Optional[
+            List[
+                Tuple[
+                    Optional[Any],
+                    Optional[Any],
+                    Optional[Any],
+                    None,
+                    Optional[Any],
+                    Optional[Any],
+                    Optional[Any],
+                ]
+            ]
+        ] = None
+        self._statement: Any = self.connection.createStatement()
+        self._result_set: Optional[Any] = None
+        self._meta_data: Optional[Any] = None
+        self._update_count: int = -1
 
     @property
-    def connection(self):
+    def connection(self) -> Any:
         return self._connection
 
     @property
-    def arraysize(self):
+    def arraysize(self) -> int:
         return self._arraysize
 
     @arraysize.setter
-    def arraysize(self, value):
+    def arraysize(self, value: int):
         if value <= 0 or value > self.DEFAULT_FETCH_SIZE:
             raise ProgrammingError(
                 "MaxResults is more than maximum allowed length {0}.".format(
@@ -49,43 +64,62 @@ class Cursor(object):
         self._arraysize = value
 
     @property
-    def rownumber(self):
+    def rownumber(self) -> Optional[int]:
         return self._rownumber
 
     @property
-    def rowcount(self):
+    def rowcount(self) -> int:
         """By default, return -1 to indicate that this is not supported."""
         return -1
 
-    @property
+    @property  # type: ignore
     @attach_thread_to_jvm
-    def has_result_set(self):
-        return self._result_set and self._meta_data and not self._result_set.isClosed()
+    def has_result_set(self) -> bool:
+        return (
+            self._result_set is not None
+            and self._meta_data is not None
+            and not self._result_set.isClosed()
+        )
 
-    @property
+    @property  # type: ignore
     @attach_thread_to_jvm
-    def description(self):
+    def description(
+        self,
+    ) -> Optional[
+        List[
+            Tuple[
+                Optional[Any],
+                Optional[Any],
+                Optional[Any],
+                None,
+                Optional[Any],
+                Optional[Any],
+                Optional[Any],
+            ]
+        ]
+    ]:
         if self._description:
             return self._description
         if not self.has_result_set:
             return None
+        meta_data = cast(Any, self._meta_data)
         self._description = [
             (
-                self._meta_data.getColumnName(i),
-                self._converter.get_jdbc_type_name(self._meta_data.getColumnType(i)),
-                self._meta_data.getColumnDisplaySize(i),
+                meta_data.getColumnName(i),
+                self._converter.get_jdbc_type_name(meta_data.getColumnType(i)),
+                meta_data.getColumnDisplaySize(i),
                 None,
-                self._meta_data.getPrecision(i),
-                self._meta_data.getScale(i),
-                self._meta_data.isNullable(i),
+                meta_data.getPrecision(i),
+                meta_data.getScale(i),
+                meta_data.isNullable(i),
             )
-            for i in xrange(1, self._meta_data.getColumnCount() + 1)
+            for i in range(1, meta_data.getColumnCount() + 1)
         ]
         return self._description
 
     @attach_thread_to_jvm
     @synchronized
-    def close(self):
+    def close(self) -> None:
         self._meta_data = None
         if self._result_set and not self._result_set.isClosed():
             self._result_set.close()
@@ -97,10 +131,10 @@ class Cursor(object):
         self._connection = None
 
     @property
-    def is_closed(self):
+    def is_closed(self) -> bool:
         return self._connection is None
 
-    def _reset_state(self):
+    def _reset_state(self) -> None:
         self._description = None
         self._result_set = None
         self._meta_data = None
@@ -108,7 +142,7 @@ class Cursor(object):
 
     @attach_thread_to_jvm
     @synchronized
-    def execute(self, operation, parameters=None):
+    def execute(self, operation: str, parameters: Optional[Dict[str, Any]] = None):
         if self.is_closed:
             raise ProgrammingError("Connection is closed.")
 
@@ -126,9 +160,12 @@ class Cursor(object):
                 self._update_count = self._statement.getUpdateCount()
         except Exception as e:
             _logger.exception("Failed to execute query.")
-            raise_from(DatabaseError(e), e)
+            raise DatabaseError(*e.args) from e
+        return self
 
-    def executemany(self, operation, seq_of_parameters):
+    def executemany(
+        self, operation: str, seq_of_parameters: List[Optional[Dict[str, Any]]]
+    ):
         for parameters in seq_of_parameters:
             self.execute(operation, parameters)
         # Operations that have result sets are not allowed with executemany.
@@ -136,7 +173,7 @@ class Cursor(object):
 
     @attach_thread_to_jvm
     @synchronized
-    def cancel(self):
+    def cancel(self) -> None:
         if self.is_closed:
             raise ProgrammingError("Connection is closed.")
         self._statement.cancel()
@@ -148,15 +185,17 @@ class Cursor(object):
         if not self.has_result_set:
             raise ProgrammingError("No result set.")
 
-        if not self._result_set.next():
+        result_set = cast(Any, self._result_set)
+        if not result_set.next():
             return None
+        if self._rownumber is None:
+            self._rownumber = 0
         self._rownumber += 1
+        meta_data = cast(Any, self._meta_data)
         return tuple(
             [
-                self._converter.convert(
-                    self._meta_data.getColumnType(i), self._result_set, i
-                )
-                for i in xrange(1, self._meta_data.getColumnCount() + 1)
+                self._converter.convert(meta_data.getColumnType(i), result_set, i)
+                for i in range(1, meta_data.getColumnCount() + 1)
             ]
         )
 
@@ -165,11 +204,11 @@ class Cursor(object):
         return self._fetch()
 
     @synchronized
-    def fetchmany(self, size=None):
+    def fetchmany(self, size: int = None):
         if not size or size <= 0:
             size = self._arraysize
         rows = []
-        for i in xrange(size):
+        for i in range(size):
             row = self._fetch()
             if row:
                 rows.append(row)
