@@ -1,82 +1,88 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, unicode_literals
-
 import datetime
+from typing import TYPE_CHECKING, FrozenSet, Type
 
 from pyathenajdbc.error import *  # noqa
 
-__version__ = "2.1.1"
-__athena_driver_version__ = "2.0.14"
+if TYPE_CHECKING:
+    from pyathenajdbc.connection import Connection
+
+__version__: str = "2.1.1"
+__athena_driver_version__: str = "2.0.15.1000"
 
 # Globals https://www.python.org/dev/peps/pep-0249/#globals
-apilevel = "2.0"
-threadsafety = 3
-paramstyle = "pyformat"
+apilevel: str = "2.0"
+threadsafety: int = 3
+paramstyle: str = "pyformat"
 
-ATHENA_JAR = "AthenaJDBC42_{0}.jar".format(__athena_driver_version__)
-ATHENA_DRIVER_DOWNLOAD_URL = (
+ATHENA_JAR: str = "AthenaJDBC42.jar"
+ATHENA_DRIVER_DOWNLOAD_URL: str = (
     "https://s3.amazonaws.com/athena-downloads/drivers/JDBC/"
-    + "SimbaAthenaJDBC_{0}/{1}".format(__athena_driver_version__, ATHENA_JAR)
+    + "SimbaAthenaJDBC-{0}/{1}".format(__athena_driver_version__, ATHENA_JAR)
 )
-ATHENA_DRIVER_CLASS_NAME = "com.simba.athena.jdbc.Driver"
-ATHENA_CONNECTION_STRING = "jdbc:awsathena://AwsRegion={region};"
-LOG4J_PROPERTIES = "log4j.properties"
+ATHENA_DRIVER_CLASS_NAME: str = "com.simba.athena.jdbc.Driver"
+ATHENA_CONNECTION_STRING: str = "jdbc:awsathena://AwsRegion={region};"
+LOG4J_PROPERTIES: str = "log4j.properties"
 
 
-class DBAPITypeObject:
+class DBAPITypeObject(FrozenSet[str]):
     """Type Objects and Constructors
 
     https://www.python.org/dev/peps/pep-0249/#type-objects-and-constructors
     """
 
-    def __init__(self, *values):
-        self.values = values
-
-    def __cmp__(self, other):
-        if other in self.values:
-            return 0
-        if other < self.values:
-            return 1
+    def __eq__(self, other: object):
+        if isinstance(other, frozenset):
+            return frozenset.__eq__(self, other)
         else:
-            return -1
+            return other in self
 
-    def __eq__(self, other):
-        return other in self.values
+    def __ne__(self, other: object):
+        if isinstance(other, frozenset):
+            return frozenset.__ne__(self, other)
+        else:
+            return other not in self
+
+    def __hash__(self):
+        return frozenset.__hash__(self)
 
 
-STRING = DBAPITypeObject(
-    "CHAR",
-    "NCHAR",
-    "VARCHAR",
-    "NVARCHAR",
-    "LONGVARCHAR",
-    "LONGNVARCHAR",
-    "ARRAY",
-    "JAVA_OBJECT",
+STRING: DBAPITypeObject = DBAPITypeObject(
+    (
+        "CHAR",
+        "NCHAR",
+        "VARCHAR",
+        "NVARCHAR",
+        "LONGVARCHAR",
+        "LONGNVARCHAR",
+        "ARRAY",
+        "JAVA_OBJECT",
+    )
 )
-BINARY = DBAPITypeObject("BINARY", "VARBINARY", "LONGVARBINARY")
-BOOLEAN = DBAPITypeObject("BOOLEAN")
-NUMBER = DBAPITypeObject(
-    "TINYINT",
-    "SMALLINT",
-    "BIGINT",
-    "INTEGER",
-    "REAL",
-    "DOUBLE",
-    "FLOAT",
-    "DECIMAL",
-    "NUMERIC",
+BINARY: DBAPITypeObject = DBAPITypeObject(("BINARY", "VARBINARY", "LONGVARBINARY"))
+BOOLEAN: DBAPITypeObject = DBAPITypeObject(("BOOLEAN",))
+NUMBER: DBAPITypeObject = DBAPITypeObject(
+    (
+        "TINYINT",
+        "SMALLINT",
+        "BIGINT",
+        "INTEGER",
+        "REAL",
+        "DOUBLE",
+        "FLOAT",
+        "DECIMAL",
+        "NUMERIC",
+    )
 )
-DATE = DBAPITypeObject("DATE")
-DATETIME = DBAPITypeObject("TIMESTAMP")
+DATE: DBAPITypeObject = DBAPITypeObject(("DATE",))
+DATETIME: DBAPITypeObject = DBAPITypeObject(("TIMESTAMP",))
+
+Date: Type[datetime.date] = datetime.date
+Time: Type[datetime.time] = datetime.time
+Timestamp: Type[datetime.datetime] = datetime.datetime
 
 
-Date = datetime.date
-Time = datetime.time
-Timestamp = datetime.datetime
-
-
-def connect(*args, **kwargs):
+def connect(*args, **kwargs) -> "Connection":
     from pyathenajdbc.connection import Connection
 
     return Connection(*args, **kwargs)
