@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import, unicode_literals
-
 import codecs
 import contextlib
 import functools
@@ -19,27 +17,33 @@ class Env(object):
         ), "Required environment variable `AWS_ATHENA_S3_STAGING_DIR` not found."
 
 
-def with_cursor(fn):
-    @functools.wraps(fn)
-    def wrapped_fn(self, *args, **kwargs):
-        with contextlib.closing(self.connect()) as conn:
-            with conn.cursor() as cursor:
-                fn(self, cursor, *args, **kwargs)
+def with_cursor(**opts):
+    def _with_cursor(fn):
+        @functools.wraps(fn)
+        def wrapped_fn(self, *args, **kwargs):
+            with contextlib.closing(self.connect(**opts)) as conn:
+                with conn.cursor() as cursor:
+                    fn(self, cursor, *args, **kwargs)
 
-    return wrapped_fn
+        return wrapped_fn
+
+    return _with_cursor
 
 
-def with_engine(fn):
-    @functools.wraps(fn)
-    def wrapped_fn(self, *args, **kwargs):
-        engine = self.create_engine()
-        try:
-            with contextlib.closing(engine.connect()) as conn:
-                fn(self, engine, conn, *args, **kwargs)
-        finally:
-            engine.dispose()
+def with_engine(**opts):
+    def _with_engine(fn):
+        @functools.wraps(fn)
+        def wrapped_fn(self, *args, **kwargs):
+            engine = self.create_engine(**opts)
+            try:
+                with contextlib.closing(engine.connect()) as conn:
+                    fn(self, engine, conn, *args, **kwargs)
+            finally:
+                engine.dispose()
 
-    return wrapped_fn
+        return wrapped_fn
+
+    return _with_engine
 
 
 def read_query(path):
